+++ pkg/testacc/data_source_functions_acceptance_test.go
@@ -4,17 +4,12 @@ package testacc
 
 import (
 	"fmt"
-	"regexp"
 	"testing"
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config/datasourcemodel"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config/model"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config/providermodel"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/testdatatypes"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/testenvs"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/testprofiles"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/provider/previewfeatures"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/provider/resources"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/tfversion"
@@ -72,56 +67,4 @@ data "snowflake_functions" "functions" {
   depends_on = [snowflake_function_java.f1, snowflake_function_java.f2]
 }
 `, functionsSetup, TestDatabaseName, TestSchemaName)
-}
-
-func TestAcc_Functions_gh3822_bcr2025_03(t *testing.T) {
-	t.Setenv(string(testenvs.ConfigureClientOnce), "")
-
-	schema, schemaCleanup := secondaryTestClient().Schema.CreateSchema(t)
-	t.Cleanup(schemaCleanup)
-
-	_, functionCleanup := secondaryTestClient().Function.CreatePythonInSchema(t, schema.ID())
-	t.Cleanup(functionCleanup)
-
-	providerModel := providermodel.SnowflakeProvider().WithProfile(testprofiles.Secondary).
-		WithPreviewFeaturesEnabled(string(previewfeatures.FunctionsDatasource))
-	functionsModel := datasourcemodel.Functions("test", schema.ID().DatabaseName(), schema.ID().Name())
-
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() { TestAccPreCheck(t) },
-		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-			tfversion.RequireAbove(tfversion.Version1_5_0),
-		},
-		Steps: []resource.TestStep{
-			{
-				ExternalProviders: ExternalProviderWithExactVersion("2.2.0"),
-				PreConfig: func() {
-					secondaryTestClient().BcrBundles.DisableBcrBundle(t, "2025_03")
-				},
-				Config: config.FromModels(t, providerModel, functionsModel),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(functionsModel.DatasourceReference(), "database", schema.ID().DatabaseName()),
-					resource.TestCheckResourceAttr(functionsModel.DatasourceReference(), "schema", schema.ID().Name()),
-					resource.TestCheckResourceAttr(functionsModel.DatasourceReference(), "functions.#", "1"),
-				),
-			},
-			{
-				ExternalProviders: ExternalProviderWithExactVersion("2.2.0"),
-				PreConfig: func() {
-					secondaryTestClient().BcrBundles.EnableBcrBundle(t, "2025_03")
-				},
-				Config:      config.FromModels(t, providerModel, functionsModel),
-				ExpectError: regexp.MustCompile("could not parse arguments"),
-			},
-			{
-				ProtoV6ProviderFactories: TestAccProtoV6ProviderFactories,
-				Config:                   config.FromModels(t, providerModel, functionsModel),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(functionsModel.DatasourceReference(), "database", schema.ID().DatabaseName()),
-					resource.TestCheckResourceAttr(functionsModel.DatasourceReference(), "schema", schema.ID().Name()),
-					resource.TestCheckResourceAttr(functionsModel.DatasourceReference(), "functions.#", "1"),
-				),
-			},
-		},
-	})
 }
